refactor: split cubecl-cpu in two
