+++ crates/cubecl-cpu/Cargo.toml
@@ -21,16 +21,22 @@ default = [
     "cubecl-runtime/default",
     "cubecl-common/default",
     "cubecl-core/default",
+    "cubecl-llvm/default",
 ]
 
-pliron-dump = []
-std = ["cubecl-runtime/std", "cubecl-common/std", "cubecl-core/std"]
+pliron-dump = ["cubecl-llvm/pliron-dump"]
+std = [
+    "cubecl-runtime/std",
+    "cubecl-common/std",
+    "cubecl-core/std",
+    "cubecl-llvm/std",
+]
 
 tracing = [
     "cubecl-runtime/tracing",
     "cubecl-common/tracing",
     "cubecl-core/tracing",
-    "cubecl-opt/tracing",
+    "cubecl-llvm/tracing",
 ]
 
 
@@ -42,9 +48,7 @@ cubecl-core = { path = "../cubecl-core", version = "=0.11.0-pre.2", default-feat
     "std",
 ] }
 cubecl-environment = { path = "../cubecl-environment", version = "=0.11.0-pre.2", default-features = false }
-cubecl-opt = { path = "../cubecl-opt", version = "=0.11.0-pre.2", default-features = false, features = [
-    "std",
-] }
+cubecl-llvm = { path = "../cubecl-llvm", version = "=0.11.0-pre.2", default-features = false }
 cubecl-runtime = { path = "../cubecl-runtime", version = "=0.11.0-pre.2", default-features = false, features = [
     "channel-mutex",
     "std",
@@ -53,22 +57,12 @@ cubecl-runtime = { path = "../cubecl-runtime", version = "=0.11.0-pre.2", defaul
 cubecl-std = { path = "../cubecl-std", version = "=0.11.0-pre.2", default-features = false }
 
 
-bytemuck = { workspace = true }
 crossbeam-utils = "0.8.21"
 derive-new = { workspace = true }
-half = { workspace = true }
 log = { workspace = true }
-paste = { workspace = true }
 pliron = { workspace = true, features = ["std"] }
-pliron-llvm = { workspace = true, features = ["llvm-sys"] }
-serde = { workspace = true }
 spin = { workspace = true, features = ["mutex", "spin_mutex"] }
 sysinfo = { workspace = true }
-# Not used directly, depended on to select its features.
-llvm-sys = { workspace = true }
-
-[build-dependencies]
-tracel-llvm-bundler = { workspace = true }
 
 [target.'cfg(any(target_os = "android", target_os = "linux", target_os = "macos", target_os = "freebsd"))'.dependencies]
 libc = { workspace = true }
@@ -83,6 +77,7 @@ cubecl-core = { path = "../cubecl-core", version = "=0.11.0-pre.2", features = [
 cubecl-std = { path = "../cubecl-std", version = "=0.11.0-pre.2", features = [
     "export_tests",
 ] }
+half = { workspace = true }
 paste = { workspace = true }
 pretty_assertions = { workspace = true }
 test-log = { workspace = true, features = ["trace"] }
