+++ crates/cubecl-cpu/src/compute/schedule.rs
@@ -1,9 +1,10 @@
-use crate::{compiler::jit::engine::PlironEngine, compute::stream::CpuStream};
+use crate::compute::stream::CpuStream;
 use cubecl_common::bytes::Bytes;
 use cubecl_core::{
     CubeDim, MemoryConfiguration, ir::MemoryDeviceProperties, server::MetadataBindingInfo,
 };
 use cubecl_environment::stream::StreamId;
+use cubecl_llvm::PlironEngine;
 use cubecl_runtime::{
     logging::ServerLogger,
     storage::{BytesResource, ManagedResource},
