+++ crates/cubecl-cpu/src/compute/threadpool/compute_task.rs
@@ -2,10 +2,9 @@ use std::sync::{Arc, atomic::AtomicU64};
 
 use crossbeam_utils::CachePadded;
 
-use crate::{
-    compiler::jit::{data::PlironData, engine::PlironEngine},
-    compute::threadpool::ThreadTask,
-};
+use cubecl_llvm::{PlironData, PlironEngine};
+
+use crate::compute::threadpool::ThreadTask;
 
 pub struct ComputeTask {
     pub pliron_engine: PlironEngine,
