+++ crates/cubecl-cpu/src/lib.rs
@@ -287,7 +287,6 @@ mod tests {
     }
 }
 
-pub mod compiler;
 pub mod compute;
 pub mod device;
 pub mod frontend;
