+++ crates/cubecl-llvm/src/shared/polyfill/synchronization.rs
@@ -15,9 +15,9 @@ use pliron::builtin::types::{IntegerType, Signedness};
 use pliron::dict_key;
 use pliron_llvm::ops as llvm;
 
-use crate::compiler::entrypoint::runtime_arg;
-use crate::compiler::polyfill::LowerOp;
-use crate::compiler::polyfill::ordered_atomic::{
+use crate::shared::entrypoint::runtime_arg;
+use crate::shared::polyfill::LowerOp;
+use crate::shared::polyfill::ordered_atomic::{
     atomic_fetch_add_acq_rel, atomic_load_acquire, atomic_store_release,
 };
